fix: lastViewed 신청 목록 조회 부분 null 초기값 문제, 수정
